pyproject.toml
[project]
name = "exebac"
version = "0.1.0"
description = ""
authors = [
    {name = "juliocouteau",email = "juliosantss5334@hotmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
